+++ CHANGES.md
@@ -1,5 +1,9 @@
-#### 1.0.1
+#### 0.2.0
+* Button colors now work with Customizer.
+* Update Hybrid-Core to be compatible with WP 4.1 changes.
+
+#### 0.1.1
 * Add [Github Updater](https://github.com/afragen/github-updater)
 
-#### 1.0
+#### 0.1.0
 * Initial release